app/backend/requirements.txt
# Azure SDK and AI Services
azure-identity                                # Azure authentication and credentials management
azure-ai-projects==1.0.0b12                         # Azure AI project management
azure-ai-inference[opentelemetry]             # Azure AI inference with OpenTelemetry support
azure-ai-evaluation                           # Azure AI model evaluation tools
azure-ai-contentsafety                        # Azure content safety services
azure-search-documents==11.7.0b1                   # Azure AI Search integration
azure-ai-ml
azure-ai-agents==1.0.0
azure-storage-blob
azure-ai-documentintelligence==1.0.0
azure-ai-voicelive==1.0.0
azure-ai-voicelive[aiohttp]==1.0.0

# Telemetry and Monitoring
azure-monitor-opentelemetry                  # Azure Monitor OpenTelemetry integration
azure-core-tracing-opentelemetry             # Azure core OpenTelemetry tracing
opentelemetry-exporter-otlp                  # OpenTelemetry protocol exporter
opentelemetry-sdk                            # OpenTelemetry SDK
opentelemetry-instrumentation-openai-v2 
opentelemetry-exporter-otlp-proto-grpc


# Jupyter and Development Tools
jupyter                                      # Jupyter notebook interface
python-dotenv                                # Environment variable management
notebook                                     # Jupyter notebook core
ipykernel                                    # IPython kernel for Jupyteron
asyncio                                      # Async I/O support
typing                                       # Type hinting support

# Black
black[jupyter]


# Realtime
aiohttp
openai
openai[realtime]
fastrtc[vad]
pyaudio


# WebSocket Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
